% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fbroc.R
\name{conf}
\alias{conf}
\title{Generic S3 function to calculate confidence regions based upon ROC curves}
\usage{
conf(roc, ...)
}
\arguments{
\item{roc}{The object for which to calculate the performance.}

\item{...}{Further arguments to perf.
#'}
}
\description{
For using this on individual ROC curves as implemented by objects of class \code{fbroc.roc} see
\link{\code{conf.fbroc.roc}}. For paired ROC curves (class \code{conf.paired.roc}) see
\link{\code{conf.fbroc.paired.roc}}.
}
\seealso{
\link{\code{conf.fbroc.roc}}, \link{\code{conf.fbroc.paired.roc}}
}

