% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbroc.R
\docType{package}
\name{fbroc-package}
\alias{fbroc-package}
\title{fbroc: Fast Algorithms to Bootstrap Receiver Operating Characteristics Curves}
\description{
Implements a very fast C++ algorithm to quickly bootstrap receiver operating characteristics (ROC) curves and derived performance metrics, including the area under the curve (AUC) and the partial area under the curve as well as the true and false positive rate. The analysis of paired receiver operating curves is supported as well, so that a comparison of two predictors is possible. You can also plot the results and calculate confidence intervals. On a typical desktop computer the time needed for the calculation of 100000 bootstrap replicates given 500 observations requires time on the order of magnitude of one second.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/erikpeter/fbroc}
  \item Report bugs at \url{https://github.com/erikpeter/fbroc/issues}
}

}
\author{
\strong{Maintainer}: Erik Peter \email{jerikpeter@googlemail.com}

Authors:
\itemize{
  \item Erik Peter \email{jerikpeter@googlemail.com}
}

}
\keyword{internal}
