% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fbroc.R
\name{perf}
\alias{perf}
\title{Generic S3 function to calculate performance estimates based upon ROC curves}
\usage{
perf(roc, ...)
}
\arguments{
\item{roc}{The object for which to calculate the performance.}

\item{...}{Further arguments to perf.}
}
\description{
For using this on individual ROC curves as implemented by objects of class \code{fbroc.roc} see
\link{\code{perf.fbroc.roc}}. For paired ROC curves (class \code{fbroc.paired.roc}) see
\link{\code{perf.fbroc.paired.roc}}.
}
\seealso{
\link{\code{perf.fbroc.roc}}, \link{\code{perf.fbroc.paired.roc}}
}

