% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbroc.paired.methods.R
\name{print.fbroc.perf.paired}
\alias{print.fbroc.perf.paired}
\title{Prints information about a \code{fbroc.perf.paired} object}
\usage{
\method{print}{fbroc.perf.paired}(x, ...)
}
\arguments{
\item{x}{Object of class \code{fbroc.perf}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Character containing the text that is also printed.
}
\description{
Prints the information about the bootstrap results for an object of class
\code{fbroc.perf.paired}. This information includes the number of bootstrap
replicates, the metric used and estimates for both the individual classifiers and the
difference in performance including confidence intervals. Finally, an estimate for the 
correlation between the performance estimates of the two classifiers is also given.
}
\seealso{
\code{\link{perf.fbroc.paired.roc}}
}
