% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbroc.methods.R
\name{print.fbroc.perf}
\alias{print.fbroc.perf}
\title{Prints information about a \code{fbroc.perf} object}
\usage{
\method{print}{fbroc.perf}(x, ...)
}
\arguments{
\item{x}{Object of class \code{fbroc.perf}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Character containing the text that is also printed.
}
\description{
Prints the information about the bootstrap results for an object of class
\code{fbroc.perf}. This information includes the number of bootstrap
replicates, the metric used and the estimate with confidence interval.
}
\seealso{
\code{\link{perf.fbroc.roc}}
}
