% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fbroc.paired.R
\name{boot.paired.roc}
\alias{boot.paired.roc}
\title{Bootstrap paired ROC curve}
\usage{
boot.paired.roc(pred1, pred2, true.class, stratify = TRUE, n.boot = 1000,
  use.cache = FALSE, tie.strategy = NULL)
}
\arguments{
\item{true.class}{A logical vector. TRUE indicates the sample belonging to the
positive class.}

\item{stratify}{Logical. Indicates whether we use stratified bootstrap.
Default to TRUE. Non-stratified bootstrap is not yet implemented.}

\item{n.boot}{A number that will be coerced to integer. Specified the
number of bootstrap replicates. Defaults to 1000.}

\item{use.cache}{If true the bootstrapping results for the
ROC curve will be pre-cached. This increases speed when the object is used often, but also
takes up more memory.}

\item{tie.strategy}{How to handle ties. See details below.}
}
\description{
Bootstrap paired ROC curve
}
\examples{
y <- rep(c(TRUE, FALSE), each = 500)
x1 <- rnorm(1000) + y
x2 <- 0.5*x1 + 0.5*rnorm(1000) + y
result.boot <- boot.paired.roc(x1, x2, y)
}
\seealso{
\url{http://www.epeter-stats.de/roc-curves-and-ties/}, \code{\link{plot.fbroc.roc}},
\code{\link{print.fbroc.roc}}
}

