% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fbroc.R
\name{boot.roc}
\alias{boot.roc}
\title{Bootstrap ROC curve}
\usage{
boot.roc(pred, true.class, stratify = TRUE, n.boot = 1000, seed = NULL)
}
\arguments{
\item{pred}{A numeric vector. Contains predictions. \code{boot.roc} assumes
that a high prediction is evidence for the observation belonging to the
positive class.}

\item{true.class}{A logical vector. TRUE indicates the sample belonging to the
positive class.}

\item{stratify}{Logical. Indicates whether we use stratified boostrap.
Default to TRUE. Non-stratified bootstrap is not yet implemented.}

\item{n.boot}{A number that will be coerced to integer. Specified the
number of bootstrap replicates. Defaults to 1000.}

\item{seed}{A number that will be coerced to integer. Used to initialise the
  random number generator used. If not specified will be set to a random number
  between 1 and 1e7.}
}
\description{
\code{boot.roc} calculates the ROC curve, initialises the settings for the
bootstrap and calculates the bootstrap results for the true and false
positive rate at every relevant threshold. Missing values are removed with
a warning prior to bootstrapping.
}
\examples{
y <- rep(c(TRUE, FALSE), each = 500)
x <- rnorm(1000) + y
result.boot <- boot.roc(x, y)
}

