% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fbroc.paired.R
\name{conf.fbroc.paired.roc}
\alias{conf.fbroc.paired.roc}
\title{BlaBla}
\usage{
\method{conf}{fbroc.paired.roc}(roc, conf.level = 0.95, conf.for = "TPR",
  steps = 250, ...)
}
\arguments{
\item{roc}{Object of class \code{fbroc.roc}.}

\item{conf.level}{Confidence level to be used for the confidence intervals.}

\item{steps}{Number of discrete steps for the FPR at which the TPR is
calculated. TPR confidence intervals are given for all FPRs in
\code{seq(0, 1, by = (1 / steps))}. Defaults to 250.}
}
\value{
A data.frame containing the FPR steps and the lower and upper bounds
of the confidence interval for the TPR.
}
\description{
Placeholder
}
\seealso{
\code{\link{boot.roc}}
}

