% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbroc.R
\name{conf}
\alias{conf}
\title{Generic S3 function to calculate confidence regions for ROC curves}
\usage{
conf(roc, ...)
}
\arguments{
\item{roc}{The object for which to calculate the performance.}

\item{...}{Further arguments to perf.}
}
\description{
For using this on individual ROC curves as implemented by objects of class \code{fbroc.roc} see 
\code{\link{conf.fbroc.roc}}. For paired ROC curves (class \code{conf.paired.roc}) see 
\code{\link{conf.fbroc.paired.roc}}.
}
\seealso{
\code{\link{conf.fbroc.roc}}, \code{\link{conf.fbroc.paired.roc}}
}
