\name{fbroc-package}
\alias{fbroc-package}
\alias{fbroc}
\docType{package}
\title{
\packageTitle{fbroc}
}
\description{
\packageDescription{fbroc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fbroc}
\packageIndices{fbroc}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{fbroc}

Maintainer: \packageMaintainer{fbroc}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
