% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fbroc.R
\docType{package}
\name{fbroc}
\alias{fbroc}
\alias{fbroc-package}
\title{fbroc: A package for fast bootstrap analysis of ROC curves}
\description{
The fbroc enables the fast boostrap analysis of ROC curves for simulation
studies and shiny applications by using a fast
algorithm where the cost of a single bootstrap replicate is \eqn{O(n)} when
used with a total of n samples. The algorithm is implemented in C++ to further
increase the efficiency. Currently 100000 bootstrap iterations for 500
observations take about one second.
}
\section{Important fbroc functions}{

\describe{
\item{\code{\link{boot.roc}}}{Use \code{boot.roc} to bootstrap a ROC curve}
\item{\code{\link{perf.roc}}}{Estimate performance and calculate confidence
intervals for a bootstrapped ROC curve}
}
}

\section{Details}{

The algorithm works by first determining the critical thresholds of the ROC
curve - cutoffs at which the curve changes directions. Each observation is linked
to the specific thresholds at which they first cause a change in the TPR
or FPR. Calculating this link and performing the bootstrap iteration on them
allows us to bootstrap the link directly and construct the bootstrap ROC
curve very quickly. Since multiple observation can be linked to the same
threshold, it is difficult to implement the algorithm efficiently in R.
Therefore \code{fbroc} implements it in C++.
\cr \cr
All bootstrap confidence interval are based on the percentile method.
}

\section{Notes}{

Package \code{fbroc} is still in an early stage. In the future it will support
more performance metrics based on ROC curves such as TPR at a fixed FPR,
partial AUC and more. In addition, support for paired ROC curves and more
sophisticated bootstrap confidence interval calculation will be added.
}
\examples{
y <- rep(c(TRUE, FALSE), each = 500)
x <- rnorm(1000) + y
result.boot <- boot.roc(x, y, n.boot = 100)
plot(result.boot)
perf.roc(result.boot, "auc")
perf.roc(result.boot, "auc", conf.level = 0.99)
}

