% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fbroc.methods.R
\name{plot.fbroc.perf}
\alias{plot.fbroc.perf}
\title{Plots performance}
\usage{
\method{plot}{fbroc.perf}(x, bins = NULL, col = "white",
  fill = "lightblue", print.plot = TRUE, show.conf = TRUE,
  conf.text = TRUE, ...)
}
\arguments{
\item{x}{Object of class \code{perf.roc} to be plotted.}

\item{bins}{Number of bins for histogram. Defaults to between 20 and 60
depending on number of bootstrap replicates.}

\item{col}{Colour of outline of histogram bars. Defaults to white.}

\item{fill}{Fill of histogram bars. Defaulst to lightblue.}

\item{print.plot}{Logical specifying whether the plot should be printed.}

\item{show.conf}{Logical specifying whether the confidence interval
should be displayed.}

\item{conf.text}{Logical specifying whether the confidence interval limits
should also be displayed as text.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A ggplot, so that the user can customize the plot further.
}
\description{
Plots performance
}
\examples{
y <- rep(c(TRUE, FALSE), each = 500)
x <- rnorm(1000) + y
result.boot <- boot.roc(x, y, n.boot = 10000)
result.perf <- perf.roc(result.boot, "auc")
plot(result.perf)
}
\seealso{
\code{\link{perf.roc}}
}

